requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.16
langchain-community==0.2.16
langchain-qdrant==0.1.4
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
arize-phoenix-otel==0.6.1
openinference-instrumentation-langchain==0.1.62
httpx==0.27.0
sentence-transformers==3.0.1
